// src/app/constants/BaseConstants.enum.ts
export enum BaseConstants {
    LOCAL_STORAGE_CURRENT_USER_NAME = 'ranjana.username',
    LOCAL_STORAGE_ACCESS_TOKEN = 'ranjana.accesstoken',
    LOCAL_STORAGE_REFRESH_TOKEN = 'ranjana.refreshtoken',
    LOCAL_STORAGE_USER_TYPE = 'ranjana.usertype',

    ONLINE_SHEDULER_TIME = 30000,// in seconds
}