// src/app/constants/APIConstants.enum.ts
export enum APIConstants {
    BASE_URL = 'http://128.199.134.180',
    API_AUTH_CLIENT_ID = 'bvvsLWc39br7yKJV5KjKTCcfdQhBs9Zza6bdVwn0',
    API_AUTH_CLIENT_SECRET = 'cgRRxskW3QjloVOthtvOKP7XsAimhLMHYhDK6tMEvUGyKhuZtTxUihAt7q97AVuHBmeKXBv7Owmz1hMP48yq4rgYuBO3xmLd90MBG1llt7fLBrs2YMcQ9b6JGiL4bhK0',
}